Created abstraction layer for HW button.
Abstracts HW specific information as a generic push button in the aws_iot_combined_demo.c file.  Provides a clean template for mapping the abstract HW push button functions into different versions of HW.
